+++ modules/cirrus/task/outputs.tf
@@ -6,6 +6,11 @@ output "lambda" {
   value = {
     function_arn = one(aws_lambda_function.task[*].arn)
     role_arn     = one(aws_iam_role.task_lambda[*].arn)
+    resolved_ecr_image_digest = (
+      local.lambda_resolve_ecr_tag_to_digest
+      ? data.aws_ecr_image.lambda_task_image[0].image_digest
+      : null
+    )
   }
 }
 
@@ -18,5 +23,10 @@ output "batch" {
     job_queue_arn      = try(local.batch_compute_config.job_queue_arn, null)
     job_definition_arn = one(aws_batch_job_definition.task[*].arn)
     role_arn           = one(aws_iam_role.task_batch[*].arn)
+    resolved_ecr_image_digest = (
+      local.batch_resolve_ecr_tag_to_digest
+      ? data.aws_ecr_image.batch_task_image[0].image_digest
+      : null
+    )
   }
 }